Make format strings correspond exactly to parameters
Subject: Re: Hereis weirdness in 5.003_26

On Thu, 13 Feb 1997 15:13:18 -0500 (EST), Chip Salzenberg <chip@atlantic.net> said:
> According to Ilya Zakharevich:
>> 
>> Why was not it catched by gcc prototypes?
> 
> I don't use gcc.  Does anyone who uses gcc compile with -Wproto ?

-Wformat, you mean.  I wasn't previously, but I will in the future.  It
turned up a few bugs (some universal, some which would trigger only
where I32 != int or the like).  I think -Wformat should be added to
CFLAGS automatically when appropriate.

p5p-msgid: <pz7mkc1h0g.fsf@eeyore.ibcinc.com>

+++ gv.c
@@ -341,7 +341,7 @@ I32 create;
 #ifdef VMS
 	warn("Weird package name \"%s\" truncated", name);
 #else
-	warn("Weird package name \"%.*s...\" truncated", namelen, name);
+	warn("Weird package name \"%.*s...\" truncated", (int)namelen, name);
 #endif
     }
     Copy(name,tmpbuf,namelen,char);
